Loïc LAMBERT
You will find attached a new version of SDL_ttf.c (and the corresponding patch) to fix underline and strikethrough according to the patch of Jon Atkins. It computes the underline/strikethrough top row and bottom row for glyphs and updates the surface size if needed.
You may prefer to change the glyph 'rows' in Find_Glyph(). I don't know enough the code of SDL_ttf to make such a modification.

I attached a PNG to see the changes I made.
The first images are a solid glyph 'A' with strikethrough and underline. The underline is not visible because the surface is not high enough. The first A is badly drawn after a regression in revision 5164. The other problem is the width of the surface: the surface width was the glyph _pitch_ for solid drawing, but it was glyph _width_ for the other. I changed that.
The other images are a text (solid), without and with the underline and strikethrough, then the drawing of the A solid, shaded, blended and blended outlined. The fonts are Vera Bold and FletcherGothicFLF size 55.

List of changes:
1- Fix drawing of solid text and glyph after revision 5164 (one 'ch' left around line 714)
2- Update surface height in TTF_RenderGlyph_*() to take underline into account
3- Fix line row for underline and strikethrough in TTF_RenderGlyph_*() - thanks to Jon Atkins
4- Change glyph surface width in TTF_RenderGlyph_Solid()